Moved routines in perturbation_auf that need to be in misc_functions to be more properly used in multiple code locations.
